+++ crates/unicode/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "rustpython-unicode"
+description = "Runtime-independent CPython-compatible Unicode semantics and data for RustPython and related Python tooling."
+edition = { workspace = true }
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+license = { workspace = true }
+rust-version = { workspace = true }
+
+[dependencies]
+rustpython-wtf8 = { workspace = true }
+
+icu_properties = { workspace = true }
+icu_normalizer = { workspace = true }
+unicode_names2 = { workspace = true }
+
+[build-dependencies]
+icu_properties = { workspace = true }
+
+[lints]
+workspace = true
