+++ crates/unicode/src/lib.rs
@@ -1,0 +1,51 @@
+//! Runtime-independent CPython-compatible Unicode semantics and data.
+//!
+//! Every entry point operates on plain `char`/`u32`/`CodePoint`/`&Wtf8` values
+//! so it can be shared by any Python runtime; argument extraction and Python
+//! exception mapping stay with the caller. There is no global mutable state and
+//! results depend only on inputs.
+
+pub mod data;
+pub mod normalize;
+
+pub use data::{Ucd, character_name, lookup_character, unicode_version};
+pub use normalize::{NormalizeForm, is_normalized, normalize};
+
+#[cfg(test)]
+mod tests {
+    use rustpython_wtf8::{CodePoint, Wtf8Buf};
+
+    use crate::{NormalizeForm, Ucd, character_name, is_normalized, lookup_character, normalize};
+
+    fn cp(ch: char) -> CodePoint {
+        CodePoint::from(ch)
+    }
+
+    #[test]
+    fn data_queries_match_unicodedata_behavior() {
+        let ucd = Ucd::new(true);
+        assert_eq!(ucd.category(cp('A')), "Lu");
+        assert_eq!(ucd.category(CodePoint::from_u32(0xD800).unwrap()), "Cs");
+        assert_eq!(lookup_character("SNOWMAN"), Some('☃'));
+        assert_eq!(character_name('☃').as_deref(), Some("SNOWMAN"));
+        assert_eq!(ucd.decimal(cp('५')), Some(5));
+        assert_eq!(ucd.digit(cp('²')), Some(2));
+        let third = ucd.numeric(cp('⅓')).unwrap();
+        assert!((third - 1.0 / 3.0).abs() < 1e-6, "got {third}");
+    }
+
+    #[test]
+    fn ucd_3_2_0_view_differs_from_modern() {
+        let legacy = Ucd::new(false);
+        assert_eq!(legacy.unidata_version(), "3.2.0");
+    }
+
+    #[test]
+    fn normalization_round_trips() {
+        let composed = Wtf8Buf::from("é");
+        let decomposed = normalize(NormalizeForm::Nfd, &composed);
+        assert_eq!(normalize(NormalizeForm::Nfc, &decomposed), composed);
+        assert!(is_normalized(NormalizeForm::Nfc, "é".as_bytes()));
+        assert!(!is_normalized(NormalizeForm::Nfd, "é".as_bytes()));
+    }
+}
