+++ crates/unicode/unicode/latest/DerivedNumericValues.txt
