+++ crates/unicode/unicode/latest/NormalizationCorrections.txt
