+++ crates/unicode/unicode/ucd32/DerivedBidiClass-3.2.0.txt
