+++ crates/unicode/unicode/ucd32/DerivedBinaryProperties-3.2.0.txt
