+++ crates/unicode/unicode/ucd32/DerivedEastAsianWidth-3.2.0.txt
