+++ crates/unicode/unicode/ucd32/DerivedGeneralCategory-3.2.0.txt
