+++ crates/unicode/unicode/ucd32/DerivedNumericType-3.2.0.txt
