+++ crates/unicode/unicode/ucd32/DerivedNumericValues-3.2.0.txt
